Restructured code with the new modular bundles: filter1 -> filter2 -> generate
