+++ build.sbt
@@ -16,10 +16,10 @@ resolvers := Seq(
 bucketSuffix  := "era7.com"
 
 libraryDependencies ++= Seq(
+  "era7bio"       %% "db-rnacentral"   % "0.5.0-SNAPSHOT",
   "ohnosequences" %% "fastarious"      % "0.6.0",
   "ohnosequences" %% "blast-api"       % "0.7.0",
   "ohnosequences" %% "statika"         % "2.0.0-M5",
-  "era7bio"       %% "db-rnacentral"   % "0.4.0",
   "ohnosequences-bundles" %% "bio4j-dist" % "0.2.0",
   // Test:
   "era7"          %% "defaults"  % "0.1.0" % Test,
